requirements.txt
streamlit
transformers
torch
newsapi-python
python-dotenv
bs4
plotly